Use a function for logging
Create a printf-style function for inserting an item into the log, also
ensuring it's visible.  Use NULL to add the last error (assumed to occur
after the log message indicating the cause of the error).  Use a flag to
indicate a log item should be replaced, setting its text directly,
rather than deleting and inserting again.